requirements.txt
asgiref==3.2.7
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.41.0
django-app-settings==0.7.0
django-appconf==1.0.4
django-baton==1.6.0
django-braces==1.14.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-resized==0.3.11
django-select2==7.2.4
django-suit-dashboard==2.0.6
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.14.3
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.3
idna==2.9
oauth2client==1.5.2
oauthlib==3.1.0
Pillow==7.1.2
protobuf==3.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==19.1.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
